Update to PHP 5.2.0

+++ Makefile
@@ -2,7 +2,7 @@ ifndef NAVISERVER
     NAVISERVER  = /usr/local/ns
 endif
 
-php_ver 	= php-5.1.4
+php_ver 	= php-5.2.0
 php_dir 	= /usr/local/php
 
 clean:
@@ -12,12 +12,14 @@ include  $(NAVISERVER)/include/Makefile.module
 
 build:
 	rm -rf $(php_ver)
-	wget -c -O /tmp/$(php_ver).tar.gz http://www.php.net/distributions/$(php_ver).tar.gz
+	if [ ! -e /tmp/$(php_ver).tar.gz ]; then wget -c -O /tmp/$(php_ver).tar.gz http://www.php.net/distributions/$(php_ver).tar.gz; fi
 	tar -xzf /tmp/$(php_ver).tar.gz
 	ln -s `pwd` $(php_ver)/sapi/naviserver
 	(cd $(php_ver) && \
+	rm -rf configure && \
 	./buildconf --force && \
 	./configure --prefix=$(php_dir) --with-config-file-path=$(php_dir)/etc \
+	--mandir=$(php_dir)/man \
 	--with-naviserver --enable-debug --enable-sockets --enable-soap --enable-calendar \
 	--with-openssl --with-zlib --with-curl --enable-dba --with-flatfile --with-gd \
 	--with-imap --with-ldap --with-ldap-sasl --with-mcrypt=/usr --with-mhash=/usr \
